renaming app dir to api
